ml/requirements.txt

numpy==2.3.4
pillow==12.0.0
opencv-python==4.10.0.84
easyocr==1.7.2

torch==2.9.0
gymnasium==0.29.1
stable-baselines3==2.7.0
transformers==4.57.1
tokenizers==0.22.1

pandas==2.3.3
scikit-learn==1.7.2
matplotlib==3.9.2
seaborn==0.13.0

pydantic==2.12.3
fastapi==0.119.1
uvicorn[standard]==0.36.0
python-multipart==0.0.6
aiofiles==23.2.1

cryptography>=43.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2